Add initial boilerplate for Bloc architecture in Flutter
This commit introduces the initial setup of the Bloc pattern in Flutter for managing state in the application. Specifically, it adds the key components of Bloc: events, states, and the Bloc itself. Furthermore, basic widget tests are introduced, as well as the Counter page for displaying the result of state changes in the UI. Relevant Bloc files are also included.